이미지 포함 삭제기능 구현

+++ src/main/resources/static/js/post.js
@@ -33,21 +33,35 @@ var main = {
 };
 
 function post_del() {
-    $.ajax({
-         type: 'DELETE',
-         url: '/api/posts/' + $('#post-id').val(),
-         success: function() {
-             Swal.fire({
-                 title: "게시글을 삭제했어요.",
-                 icon: "success"
-             })
-             .then(() =>{
-                 window.location.href= '/board/' + $('#board').val();
-             });
-         },
-         error: function(e) {
-             alert('몰라');
-         }
+    var board = $('#board').val();
+
+    Swal.fire({
+        title: '게시글을 삭제하시나요?',
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: '삭제',
+        cancelButtonText: '취소'
+    }).then((result) => {
+        if (result.isConfirmed) {
+            $.ajax({
+                 type: 'DELETE',
+                 url: '/api/posts/' + $('#post-id').val(),
+                 success: function() {
+                     Swal.fire({
+                         title: "게시글을 삭제했어요.",
+                         icon: "success"
+                     })
+                     .then(() =>{
+                         window.location.href= '/board/' + board + '?page=0';
+                     });
+                 },
+                 error: function(e) {
+                     alert('몰라');
+                 }
+            });
+        }
     });
 }
